requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pymysql==1.1.1
pandas==2.2.2
plotly==5.24.1
reportlab==4.2.5
